+++ spec/commands/overview_spec.rb
@@ -1,31 +1,10 @@
 # frozen_string_literal: true
 
 require 'commands'
-require 'alces_utils'
-
-RSpec.shared_context 'mock overview namespaces' do
-  include AlcesUtils
-  let :config_value { 'config_value' }
-  let :static { 'static' }
-  let :fields do
-    [
-      { header: 'heading1', value: static },
-      { header: 'heading2', value: '<%= scope.config.key %>' },
-      { header: 'heading3', value: '' },
-      { header: 'missing-value' },
-      { value: 'missing-header' }
-    ]
-  end
-
-  AlcesUtils.mock self, :each do
-    ['group1', 'group2', 'group3'].map do |group|
-      config(mock_group(group), key: config_value)
-    end
-  end
-end
+require 'fixtures/shared_context/overview'
 
 RSpec.describe Metalware::Commands::Overview do
-  include_context 'mock overview namespaces'
+  include_context 'overview context'
 
   let :name_hash { { header: 'Group Name', value: '<%= group.name %>' } }
 
@@ -36,12 +15,12 @@ RSpec.describe Metalware::Commands::Overview do
   end
 
   before :each do
-    allow(Metalware::Commands::Overview::Table).to \
+    allow(Metalware::Overview::Table).to \
       receive(:new).with(any_args).and_call_original
   end
 
   def expect_table_with(*inputs)
-    expect(Metalware::Commands::Overview::Table).to \
+    expect(Metalware::Overview::Table).to \
       receive(:new).once.with(*inputs).and_call_original
   end
 
@@ -79,40 +58,6 @@ RSpec.describe Metalware::Commands::Overview do
       expect_table_with [alces.domain], overview_hash[:domain]
       run_command
     end
-  end
-end
-
-RSpec.describe Metalware::Commands::Overview::Table do
-  include_context 'mock overview namespaces'
-
-  let :namespaces { alces.groups }
-
-  let :table do
-    Metalware::Commands::Overview::Table.new(namespaces, fields).render
-  end
-
-  def header
-    table.lines[1]
-  end
-
-  def body
-    table.lines[3..-2].join("\n")
-  end
-
-  let :headers { fields.map { |h| h[:header] } }
-
-  it 'includes the headers in the table' do
-    headers.each do |h|
-      expect(header).to include(h) unless h.nil?
-    end
-  end
-
-  it 'includes the static value in the table' do
-    expect(body).to include(static)
-  end
-
-  it 'renders the values' do
-    expect(body).to include(config_value)
   end
 end
