+++ src/commands/overview.rb
@@ -23,44 +23,11 @@
 #==============================================================================
 
 require 'terminal-table'
+require 'overview/table'
 
 module Metalware
   module Commands
     class Overview < CommandHelpers::BaseCommand
-      class Table
-        attr_reader :fields
-        attr_reader :namespaces
-
-        def initialize(namespaces, fields)
-          @fields = fields
-          @namespaces = namespaces
-        end
-
-        def render
-          Terminal::Table.new(headings: headers, rows: rows).render
-        end
-
-        private
-
-        def headers
-          fields.map { |f| f[:header] }
-        end
-
-        def unrendered_values
-          fields.map { |f| f[:value] || '' }
-        end
-
-        def rows
-          namespaces.map { |namespace| row(namespace) }
-        end
-
-        def row(namespace)
-          unrendered_values.map do |value|
-            namespace.render_erb_template(value)
-          end
-        end
-      end
-
       private
 
       attr_reader :overview_data
@@ -76,14 +43,14 @@ module Metalware
 
       def print_domain_table
         fields = overview_data[:domain] || []
-        puts Table.new([alces.domain], fields).render
+        puts Metalware::Overview::Table.new([alces.domain], fields).render
       end
 
       def print_groups_table
         fields_from_yaml = overview_data[:group] || []
         name_field = { header: 'Group Name', value: '<%= group.name %>' }
         fields = [name_field].concat fields_from_yaml
-        puts Table.new(alces.groups, fields).render
+        puts Metalware::Overview::Table.new(alces.groups, fields).render
       end
     end
   end
